minor improvements to wording
TBH, I'm not a fan of the space-wasting "you must scroll and scroll to see more content" style, which was super-hip around 2015 but looks a bit dated now. But I realize that changing the style is probably out of scope, so I just made some minor text improvements.

One thing that I addressed is that the features menu points to https://linkml.io/#features, but nothing happens -- it just shows the main page. I think it's meant to bring you back up to the top if you were further down the page -- but since the menu doesn't stay fixed at the top of the page, it's no longer visible to click on if you've scrolled down, so it seems pointless. I therefore am suggesting getting rid of that menu item, but if there's somewhere useful it could point, please lmk where that is!